feat(print-this-page-functionality): remove button from VorabcheckPage, move print url to root, add test for print case
