+++ app/routes/shared/components/VorabcheckPage.tsx
@@ -1,8 +1,5 @@
-import FileDownloadIcon from "@digitalservicebund/icons/FileDownload";
 import { useLoaderData } from "@remix-run/react";
-import { useEffect } from "react";
 import Background from "~/components/Background";
-import Button from "~/components/Button";
 import Container from "~/components/Container";
 import { ProgressBar } from "~/components/form/ProgressBar";
 import ValidatedFlowForm from "~/components/form/ValidatedFlowForm";
@@ -17,15 +14,7 @@ export function VorabcheckPage() {
     formElements,
     progressProps,
     buttonNavigationProps,
-    shouldPrint,
   } = useLoaderData<typeof loader>();
-
-  useEffect(() => {
-    if (shouldPrint) {
-      window.print();
-      window.close();
-    }
-  }, [shouldPrint]);
 
   return (
     <Background backgroundColor="blue">
@@ -38,12 +27,6 @@ export function VorabcheckPage() {
                 content={contentElements}
                 className="ds-stack ds-stack-16"
                 fullScreen={false}
-              />
-              <Button
-                iconLeft={<FileDownloadIcon />}
-                text="Anleitung Herunterladen"
-                size="large"
-                href="/kontopfaendung/wegweiser/sozialleistungen-einmalzahlung?print"
               />
               <ValidatedFlowForm
                 stepData={stepData}
